Completed intial separation of manager and renderer
